KZOO-108: Do not decide BEAM process based on core count
SMP does not only depend on the numbers of cores on the running system, but it
also depends on the flags passed to the Erlang VM when starting it so let us
just get the pids for both `beam` and `beam.smp` processes and let the code
filter/search (grep) for the desired VM as it is already doing.

Prior to this change kazoo-[applications, ecallmgr] executables were failing
when running on single core systems because it seems Erlang comes with SMP
enabled by default and we don't check how many cores does the system has when
starting our VMs so they are running with SMP enabled even on single core machines
and that will brake the aforementioned executables because they would try to find
pid for `beam` processes but again since they have SMP enabled they are `beam.smp`,
thus, they will not be found and they will just reply: `DESIRED_APP is not running`.