Fix `sbin` scripts for Grep v3+ (#22)
It seems when using Grep v3+ (not exactly sure on the breaking change
version), that the `NUL` characters separating args in
`/proc/$PID/cmdline` don't match the `[^\-]` character class. This means
that the `beam` PIDs can't be found. This PR uses `tr` to replace the
`NUL` characters with spaces so the `grep` succeeds. Tested
backwards-compatibility with Grep v2.

Also DRY'd up the PID find with a shared function and ran these changes
thru ShellCheck.